put main types in classes
