+++ airbyte-integrations/base-python/.dockerignore
@@ -1,7 +1,4 @@
 *
 !Dockerfile
 !base.py
-!destination.py
-!integration.py
-!source.py
-!setup.py
+!airbyte_protocol
