+++ airbyte-integrations/base-python/airbyte_protocol/setup.py
@@ -1,0 +1,10 @@
+from setuptools import setup
+
+setup(
+    name='airbyte_protocol',
+    description='Contains classes representing the schema of the Airbyte protocol.',
+    author='Airbyte',
+    author_email='contact@airbyte.io',
+    packages=['airbyte_protocol'],
+    install_requires=[]
+)
