--- airbyte-integrations/base-python/destination.py
@@ -1,6 +1,0 @@
-from integration import Integration
-
-
-class Destination(Integration):
-    def __init__(self):
-        pass
